Add main class for GUI
