+++ src/main/java/duke/command/ExitCommand.java
@@ -20,7 +20,6 @@ public class ExitCommand extends Command {
      */
     @Override
     public String execute(TaskList tasks, Ui ui, Storage storage) {
-//        ui.exitMessage();
         return "Bye. Hope to see you again soon.";
     }
